+++ config/config.exs
@@ -1,6 +1,6 @@
 import Config
 
-config :servox,
-  native_module: Servox.Native,
-  screenshot_module: Servox.Screenshot,
+config :browse_servo,
+  native_module: BrowseServo.Native,
+  screenshot_module: BrowseServo.Screenshot,
   pools: []